+++ src/metadata/frequentslots.h
@@ -11,7 +11,7 @@ class FrequentSlots {
  public:
   FrequentSlots() {}
 
-  void allocate(uint32_t fp_hash)
+  void allocate(uint64_t fp_hash)
   {
     for (auto &reverse_mapping : _slots) {
       if (reverse_mapping.get() == nullptr) {
@@ -26,7 +26,7 @@ class FrequentSlots {
     _slots.push_back(std::move(reverse_mapping));
   }
 
-  bool query(uint32_t fp_hash, uint64_t lba)
+  bool query(uint64_t fp_hash, uint64_t lba)
   {
     for (auto &reverse_mapping : _slots) {
       if (reverse_mapping.get() == nullptr) continue;
@@ -41,7 +41,7 @@ class FrequentSlots {
     return false;
   }
 
-  void add(uint32_t fp_hash, uint64_t lba) {
+  void add(uint64_t fp_hash, uint64_t lba) {
     for (auto &reverse_mapping : _slots) {
       if (reverse_mapping.get() == nullptr) continue;
       if (reverse_mapping->_fp_hash == fp_hash) {
@@ -50,7 +50,7 @@ class FrequentSlots {
     }
   }
 
-  void remove(uint32_t fp_hash) {
+  void remove(uint64_t fp_hash) {
     for (auto &reverse_mapping : _slots) {
       if (reverse_mapping.get() == nullptr) continue;
       if (reverse_mapping->_fp_hash == fp_hash) {
@@ -61,7 +61,7 @@ class FrequentSlots {
 
  private:
   struct ReverseMapping {
-    uint32_t _fp_hash;
+    uint64_t _fp_hash;
     std::vector<uint64_t> _lbas;
   };
   std::vector<std::unique_ptr<ReverseMapping>> _slots;
